The svd_inv_thresh is not good enougth
[Problem]
The sampling of torsion motions for glycine is wrong.
But, the program shows the sampling geometries is converged.

[Cause]
Pulay et al. proposed the threshold 1e-7 is good for svd method in calculating eigenvalues of G.
Thus, we apply the corresponding threshold of 3.16e-4 for svd inversion.
However, It would fail in sampling torsion motions in UMVT method.
If you use a more tight threshold like 1e-10, the problem will solve in this case.

[Solution]
Use Numpy package to calculate B matrix inversion.
If LinAlgError("SVD did not converge") appears, the ape.log would give warning message.
Program will switch to use the svd_inv function with 3.16e-4.
If you find the sampling is not good, just use tighter thresh.