Update UMVT code
In the previous code, when torsion motion sampling terminates as cut-off energy met, the directional vector used in vibrational sampling is still the torsional direction used in torsion sampling. It might cause some problems in simulating. Thus, the modification made here is to use the projected hessian to decide the directional vectors of those torsional motions which terminate due to over the cut-off energy.